+++ app/src/main/java/com/cs246/EzBudget/mBackGrounds/BackGroundBalance.java
@@ -22,7 +22,7 @@ public class BackGroundBalance extends AsyncTask<Void,BalanceData,Void> {
     private Context context;
     private RecyclerBalanceAdapter myAdapter;
     private ArrayList<BalanceData> myBalanceData;
-    //private CategoryShowFragment teste;
+    //private DispCategoryDialogFrag teste;
     public final static int BAL_ALL = 1;
     public final static int BAL_BILLS = 2;
     public final static int BAL_INCOMES = 3;
@@ -57,7 +57,7 @@ public class BackGroundBalance extends AsyncTask<Void,BalanceData,Void> {
         Double theValue;
         String theDescription;
         String theDueDate;
-        Integer catID;
+        Long catID;
         Cursor cursor;
         switch (myConsultType) {
             case BAL_ALL:  cursor =  theBalanceData.getAllCursor();
@@ -90,7 +90,7 @@ public class BackGroundBalance extends AsyncTask<Void,BalanceData,Void> {
                 theValue = cursor.getDouble(4);
                 theDescription = cursor.getString(cursor.getColumnIndex(BalanceData.BALANCEDATA_COLUMN_DESCRIPTION));
                 theDueDate = cursor.getString(cursor.getColumnIndex(BalanceData.BALANCEDATA_COLUMN_DUE_DATE));
-                catID = cursor.getInt(cursor.getColumnIndex(BalanceData.BALANCEDATA_COLUMN_CATEGORY));
+                catID = cursor.getLong(cursor.getColumnIndex(BalanceData.BALANCEDATA_COLUMN_CATEGORY));
                 BalanceData theData = new BalanceData();
                 theData.setID(theID);
                 theData.setValue(theValue);
